requirements.txt
redis>=4.5.4
pyyaml>=6.0
prometheus_client>=0.16.0
aiohttp>=3.8.4
asyncio-mqtt>=0.13.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
aioresponses>=0.7.4